+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+v5.1.12
+----------
+ * Drop ActionLog model
+ * Switch to new editor as the default, use v1.6.10
+ * Add query field to FlowStart
+
+v5.1.11
+----------
+ * Add FlowSession.uuid which is nullable for now
+ * Update to floweditor 1.6.9, scrolling rules
+
 v5.1.10
 ----------
  * Update to flow editor 1.6.8, add completion config